+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `QEntitiesParseOptions::max_key_length()` for limiting the maximum length of keys while parsing.
 - `QEntitiesParseOptions::max_value_length()` for limiting the maximum length of values while parsing.
 - `QEntitiesParseOptions::max_entities()` for limiting the maximum number of entities while parsing.
+- `QEntitiesParseOptions::max_entity_key_values()` for limiting the maximum number of key-value pairs an entity can have while parsing.
 - `PartialEq` and `Eq` trait implementations too `QEntitiesParserLocation`.
 
 ## [0.2.2] - 2023-08-08